% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_between_centroids.R
\name{angle_between_centroids}
\alias{angle_between_centroids}
\title{Calculate the angle between groups}
\usage{
angle_between_centroids(df)
}
\arguments{
\item{df}{A data frame with specimen grouping variable (as character string) in the first column, and with PCA scores in the rest of the columns.}
}
\value{
A table of the angles between any two group centroids in the sample and a heat map of the values in the table.
}
\description{
Calculates the angle between the centres of two or more groups of specimens.
}
\examples{
angles <- angle_between_centroids(spec_info)
}
