% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asymmetric_mobility_cpp}
\alias{asymmetric_mobility_cpp}
\title{asymmetric_mobility_cpp}
\usage{
asymmetric_mobility_cpp(
  seiiar_pop,
  mobility_matrix,
  seed_matrix,
  betas,
  a,
  gamma,
  asymptomaticProb,
  asymptomaticRelativeInfectiousness,
  N = 1L,
  M = 56L
)
}
\arguments{
\item{seiiar_pop}{Data.frame}

\item{mobility_matrix}{List of data.frames}

\item{seed_matrix}{matrix of seeding cases per date (row) per geographical location (column)}

\item{betas}{Vector of floats, infection parameter, 0.6}

\item{a}{Float, 1/latent period, 1/1.9}

\item{gamma}{Float, 1/infectious period, 1/3}

\item{asymptomaticProb}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomaticRelativeInfectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of repetitions}

\item{M}{Int, Number of days}
}
\description{
Raw CPP function. Should not be called directly.
}
