% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{asymmetric_mobility_dummy_betas}
\alias{asymmetric_mobility_dummy_betas}
\title{Fake betas for asymmetric mobility}
\format{An object of class \code{numeric} of length 20.}
\usage{
asymmetric_mobility_dummy_betas
}
\description{
A vector of 20 betas (1 for each time period)
}
\keyword{datasets}
