% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{asymmetric_mobility_dummy_mobility_matrix}
\alias{asymmetric_mobility_dummy_mobility_matrix}
\title{Fake mobility matrixes for asymmetric mobility}
\format{\describe{
\item{from}{Location code.}
\item{to}{Location code.}
\item{n}{Number of people.}
}}
\usage{
asymmetric_mobility_dummy_mobility_matrix
}
\description{
A list of 20 matrices (1 for each time period)
}
\keyword{datasets}
