% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{asymmetric_mobility_dummy_seiiar_pop}
\alias{asymmetric_mobility_dummy_seiiar_pop}
\title{A fake population for asymmetric mobility}
\format{\describe{
\item{location_code}{Location code.}
\item{S}{Number of susceptible people.}
\item{E}{Number of exposed people.}
\item{I}{Number of infectious and symptomatic people.}
\item{Ia}{Number of infectious and asymptomatic people.}
\item{R}{Number of recovered people.}
}}
\usage{
asymmetric_mobility_dummy_seiiar_pop
}
\description{
A fake population for asymmetric mobility
}
\keyword{datasets}
