% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_mobility_se1e2iiar.R
\name{asymmetric_mobility_se1e2iiar}
\alias{asymmetric_mobility_se1e2iiar}
\title{asymmetric mobility SE1E2IIaR}
\usage{

  asymmetric_mobility_se1e2iiar(se1e2iiar_pop = spread::asymmetric_mobility_dummy_se1e2iiar_pop,
  mobility_matrix = spread::asymmetric_mobility_dummy_se1e2iiar_mobility_matrix,
  dynamic_seeds = NULL,
  betas = spread::asymmetric_mobility_dummy_se1e2iiar_betas,
  latent_period = 2, presymptomatic_period = 3,
  infectious_period = 5, presymptomatic_relative_infectiousness = 1.25,
  asymptomatic_prob = 0.4, asymptomatic_relative_infectiousness = 0.5,
  N = 1)
}
\arguments{
\item{se1e2iiar_pop}{Data frame containing `location_code`, `S`, `E1`, `E2`, `I`, `Ia`, and `R` for the entire population}

\item{mobility_matrix}{List (1 entry for each time period) with each entry containing a data frame with `from`, `to`, `n` for the number of people who travel. Each data frame must be complete.}

\item{dynamic_seeds}{Data.table containing `location_code`, `day`, and `n` for dynamic seeding (or `NULL`)}

\item{betas}{Vector (1 entry for each time period). Float, infection parameter, 0.6}

\item{latent_period}{Float, 2.0}

\item{presymptomatic_period}{Float 3.0}

\item{infectious_period}{Float, 5.0}

\item{presymptomatic_relative_infectiousness}{Float, relative infectiousness of presymptomatic}

\item{asymptomatic_prob}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomatic_relative_infectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of internal simulations (average taken). This is generally used for parameter fitting.}
}
\description{
For more information, look at \code{vignette("commuter_model","spread")}
}
\examples{
spread::asymmetric_mobility_se1e2iiar(
  se1e2iiar_pop = spread::asymmetric_mobility_se1e2iiar_dummy_se1e2iiar_pop,
  mobility_matrix = spread::asymmetric_mobility_se1e2iiar_dummy_mobility_matrix,
  dynamic_seeds = spread::asymmetric_mobility_se1e2iiar_dummy_dynamic_seeds,
  betas = spread::asymmetric_mobility_se1e2iiar_dummy_betas,
  latent_period = 2.0,
  presymptomatic_period = 3.0,
  infectious_period = 5.0,
  presymptomatic_relative_infectiousness = 1,
  asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0,
  N = 1
)
}
