% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_mobility_se1e2iiar_2strains.R
\name{asymmetric_mobility_se1e2iiar_2strains}
\alias{asymmetric_mobility_se1e2iiar_2strains}
\title{asymmetric mobility SE1E2IIaR two strains}
\usage{
asymmetric_mobility_se1e2iiar_2strains(
 
    se1e2iiar_2strains_pop = spread::asymmetric_mobility_se1e2iiar_dummy_se1e2iiar_2strains_pop,
  mobility_matrix = spread::asymmetric_mobility_se1e2iiar_dummy_mobility_matrix,
  dynamic_seeds = NULL,
  betas = spread::asymmetric_mobility_se1e2iiar_dummy_betas,
  inputSeed = as.numeric(Sys.time()),
  latent_period = 3,
  presymptomatic_period = 2,
  infectious_period = 5,
  b_relative_infectiousness = 1.5,
  presymptomatic_relative_infectiousness = 1.25,
  asymptomatic_prob = 0.4,
  asymptomatic_relative_infectiousness = 0.5,
  N = 1
)
}
\arguments{
\item{se1e2iiar_2strains_pop}{Data frame containing `location_code`, `S`, `E1`, `E2`, `I`, `Ia`, `R`, `E1_b`, `E2_b`, `I_b`, and`Ia_b` for the entire population}

\item{mobility_matrix}{List (1 entry for each time period) with each entry containing a data frame with `from`, `to`, `n` for the number of people who travel. Each data frame must be complete.}

\item{dynamic_seeds}{Data.table containing `location_code`, `day`, `n`, `n_b` for dynamic seeding (or `NULL`)}

\item{betas}{Data table with `locaton code`, `day`, `time`, `beta` (1 entry for each location for each time period). Float, infection parameter, 0.6}

\item{inputSeed}{Integer with seed for cpp code}

\item{latent_period}{Float, 3.0}

\item{presymptomatic_period}{Float 2.0}

\item{infectious_period}{Float, 5.0}

\item{b_relative_infectiousness}{Float, relative infectiousness of strain B}

\item{presymptomatic_relative_infectiousness}{Float, relative infectiousness of presymptomatic}

\item{asymptomatic_prob}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomatic_relative_infectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of internal simulations (average taken). This is generally used for parameter fitting.}
}
\value{
A data.table containing the following variables:
\describe{
\item{location_code}{Location code}
\item{week}{Week number}
\item{day}{Day number}
\item{time}{Time of reporting (23:59)}
\item{b_S}{Susceptibles belonging to location code}
\item{b_E1}{E1s belonging to location code}
\item{b_E2}{E2s belonging to location code}
\item{b_I}{Infectious (symptomatic) belonging to location code}
\item{b_Ia}{Infectious (asymptomatic) belonging to location code}
\item{b_R}{Recovered belonging to location code}
\item{b_E1_b}{E1_bs belonging to location code}
\item{b_E2_b}{E2_bs belonging to location code}
\item{b_I_b}{Infectious (symptomatic) strain B belonging to location code}
\item{b_Ia_b}{Infectious (asymptomatic) strain B belonging to location code}
\item{c_S}{Susceptibles currently in this location code}
\item{c_E1}{E1s currently in this location code}
\item{c_E2}{E2s currently in this location code}
\item{c_I}{Infectious (symptomatic) currently in this location code}
\item{c_Ia}{Infectious (asymptomatic) currently in this location code}
\item{c_R}{Recovered currently in this location code}
\item{c_E1_b}{E1_bs currently in this location code}
\item{c_E2_b}{E2_bs currently in this location code}
\item{c_I_b}{Infectious (symptomatic) strain B currently in this location code}
\item{c_Ia_b}{Infectious (asymptomatic) strain B currently in this location code}
\item{c_symp_incidence_a}{Transition from E2 to I currently in this location code}
\item{c_asymp_incidence_a}{Transition from E1 to Ia currently in this location code}
\item{c_symp_incidence_b}{Transition from E2_b to I_b currently in this location code}
\item{c_asymp_incidence_b}{Transition from E1_b to Ia_b currently in this location code}
\item{c_symp_incidence}{Transition from E2 and E2_b to I and I_b currently in this location code}
\item{c_asymp_incidence}{Transition from E1 and E1_b to Ia and Ia_b currently in this location code}
}
}
\description{
For more information, look at \code{vignette("commuter_model","spread")}
}
\examples{
spread::asymmetric_mobility_se1e2iiar_2strains(
  se1e2iiar_2strains_pop = spread::asymmetric_mobility_se1e2iiar_dummy_se1e2iiar_2strains_pop,
  mobility_matrix = spread::asymmetric_mobility_se1e2iiar_dummy_mobility_matrix,
  dynamic_seeds = spread::asymmetric_mobility_se1e2iiar_2strains_dummy_dynamic_seeds,
  betas = spread::asymmetric_mobility_se1e2iiar_dummy_betas,
  inputSeed = 123,
  latent_period = 3.0,
  presymptomatic_period = 2.0,
  infectious_period = 5.0,
  b_relative_infectiousness = 1.5,
  presymptomatic_relative_infectiousness = 1,
  asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0,
  N = 1
)
}
