% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asymmetric_mobility_se1e2iiar_2strains_cpp}
\alias{asymmetric_mobility_se1e2iiar_2strains_cpp}
\title{asymmetric_mobility_se1e2iiar_2strains_cpp}
\usage{
asymmetric_mobility_se1e2iiar_2strains_cpp(
  se1e2iiar_2strains_pop,
  mobility_matrix,
  seed_matrix,
  seed_matrix_b,
  betas,
  inputSeed,
  a1,
  a2,
  gamma,
  relativeInfectiousnessB,
  presymptomaticRelativeInfectiousness,
  asymptomaticProb,
  asymptomaticRelativeInfectiousness,
  N = 1L,
  M = 56L
)
}
\arguments{
\item{se1e2iiar_2strains_pop}{Data.frame}

\item{mobility_matrix}{List of data.frames}

\item{seed_matrix}{matrix of seeding cases of strain a per date per geographical location}

\item{seed_matrix_b}{matrix of seeding cases of strain b per date per geographical location}

\item{betas}{matrix of floats, number of time intervals times number of locations, infection parameter, 0.6}

\item{inputSeed}{Integer, input seed}

\item{a1}{Float, 1/latent period, 1/2.0}

\item{a2}{Float, 1/presymptomatic period, 1/3.0}

\item{gamma}{Float, 1/infectious period, 1/5.0}

\item{relativeInfectiousnessB}{Float, Relative infectiousness of strain B}

\item{presymptomaticRelativeInfectiousness}{Float, Relative infectiousness of presymptomatic infectious}

\item{asymptomaticProb}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomaticRelativeInfectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of repetitions}

\item{M}{Int, Number of days}
}
\description{
Raw CPP function. Should not be called directly.
}
