% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{asymmetric_mobility_se1e2iiar_dummy_se1e2iiar_pop}
\alias{asymmetric_mobility_se1e2iiar_dummy_se1e2iiar_pop}
\title{A fake population for asymmetric mobility se1e2iiar}
\format{\describe{
\item{location_code}{Location code.}
\item{S}{Number of susceptible people.}
\item{E1}{Number of exposed, asymptomatic people, not infectious.}
\item{E2}{Number of exposed, presymptomatic people, infectious.}
\item{I}{Number of infectious and symptomatic people.}
\item{Ia}{Number of infectious and asymptomatic people.}
\item{R}{Number of recovered people.}
}}
\usage{
asymmetric_mobility_se1e2iiar_dummy_se1e2iiar_pop
}
\description{
A fake population for asymmetric mobility se1e2iiar
}
\keyword{datasets}
