% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_process.R
\name{branching_process}
\alias{branching_process}
\title{Samples from a branching process model with a constant reproduction number}
\usage{
branching_process(initial_cases = 1, R0 = 3, dispersion = 1,
  serial_interval = 0, end_time = 10, N = 1000)
}
\description{
R0 - basic reproduction number
serial_interval a distcrete object
}
