% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{commuter_cpp}
\alias{commuter_cpp}
\title{commuter}
\usage{
commuter_cpp(pop_wo_com, di_edge_list, start_points, beta, a, gamma,
  asymptomaticProb, asymptomaticRelativeInfectiousness, N = 1L,
  M = 56L)
}
\arguments{
\item{pop_wo_com}{Data frame}

\item{di_edge_list}{Data frame}

\item{start_points}{Data frame}

\item{beta}{Float, infection parameter, 0.6}

\item{a}{Float, 1/latent period, 1/1.9}

\item{gamma}{Float, 1/infectious period, 1/3}

\item{asymptomaticProb}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomaticRelativeInfectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of repetitions}

\item{M}{Int, Number of days}
}
\description{
commuter
}
