% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_from_origin.R
\name{distance_from_origin}
\alias{distance_from_origin}
\title{Distance from origin}
\usage{
distance_from_origin(data, bs)
}
\arguments{
\item{data}{A data frame with specimens in rows; first column with group assignment; next columns with ordination scores.}

\item{bs}{Number of bootstrap samples}
}
\value{
A box plot comparing the bootstrapped estimates of the distance from the origin of the centroids of each group; a one-way ANOVA result.
}
\description{
Estimate the distance between the centroids of each group of specimens and the origin of the axes system. The coordinates of the centroid are calculated by sampling with replacement 80% of the observations.
}
\examples{
distances <- distance_from_origin(spec_info, 100);
}
