% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nor_commuters_2017_b2020}
\alias{nor_commuters_2017_b2020}
\title{Daily number of commuters from/to municipalities in Norway in 2017 (2020 borders)}
\format{
\describe{
\item{from}{Location code.}
\item{to}{Location name.}
\item{n}{Number of people.}
}
}
\source{
\url{https://www.ssb.no/statbank/table/03321}
}
\usage{
nor_commuters_2017_b2020
}
\description{
Daily number of commuters from/to municipalities in Norway in 2017 (2020 borders)
}
\keyword{datasets}
