% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{norway_commuters_2017}
\alias{norway_commuters_2017}
\title{Daily number of commuters from/to municipalities in Norway in 2017}
\format{\describe{
\item{from}{Location code.}
\item{to}{Location name.}
\item{n}{Number of people.}
}}
\source{
\url{https://www.ssb.no/statbank/table/03321}
}
\usage{
norway_commuters_2017
}
\description{
Daily number of commuters from/to municipalities in Norway in 2017
}
\keyword{datasets}
