% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{norway_pop_wo_com_2017}
\alias{norway_pop_wo_com_2017}
\title{Names of areas in Norway that previously existed.}
\format{\describe{
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_pop_wo_com_2017
}
\description{
Names of areas in Norway that previously existed.
}
\keyword{datasets}
