% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{norway_seiiar_measles_oslo_2017_b2020}
\alias{norway_seiiar_measles_oslo_2017_b2020}
\title{SEIIaR data.frame for Norway with 10 people infected in Oslo and real measles susceptibility (2020 borders)}
\format{\describe{
\item{location_code}{Location code.}
\item{S}{Number of susceptible people.}
\item{E}{Number of exposed people.}
\item{I}{Number of infectious and symptomatic people.}
\item{Ia}{Number of infectious and asymptomatic people.}
\item{R}{Number of recovered people.}
}}
\source{
\url{http://khs.fhi.no/webview/}
}
\usage{
norway_seiiar_measles_oslo_2017_b2020
}
\description{
Measles vaccination coverate rates for 16 year olds in the 5 year average
from 2014 to 2018 were used as the proportion of recovered people.
}
\keyword{datasets}
