% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{norway_seiiar_noinfected_2017_b2020}
\alias{norway_seiiar_noinfected_2017_b2020}
\title{SEIIaR data.frame for Norway with no one infected and everyone susceptible (2020 borders)}
\format{\describe{
\item{location_code}{Location code.}
\item{S}{Number of susceptible people.}
\item{E}{Number of exposed people.}
\item{I}{Number of infectious and symptomatic people.}
\item{Ia}{Number of infectious and asymptomatic people.}
\item{R}{Number of recovered people.}
}}
\usage{
norway_seiiar_noinfected_2017_b2020
}
\description{
SEIIaR data.frame for Norway with no one infected and everyone susceptible (2020 borders)
}
\keyword{datasets}
