% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{norway_seiiar_oslo_2017}
\alias{norway_seiiar_oslo_2017}
\title{Names of areas in Norway that currently exist.}
\format{\describe{
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}}
\source{
\url{https://snl.no/kommunenummer}
}
\usage{
norway_seiiar_oslo_2017
}
\description{
Names of areas in Norway that currently exist.
}
\keyword{datasets}
