% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OptimFunction}
\alias{OptimFunction}
\title{OptimFunction}
\usage{
OptimFunction(param = rep(0.75, 5), regions, d, s, startWeek)
}
\arguments{
\item{param}{a}

\item{regions}{a}

\item{d}{a}

\item{s}{a}

\item{startWeek}{a}
}
\description{
OptimFunction
}
