% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_process.R
\name{plot_quantiles_bp}
\alias{plot_quantiles_bp}
\title{plot_quantiles}
\usage{
plot_quantiles_bp(da, x = NULL, max_v = NULL, min = 0)
}
\arguments{
\item{da}{a}

\item{x}{a}

\item{max_v}{a}

\item{min}{a}
}
\description{
plot_quantiles
}
