% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_dist_group.R
\name{proc_dist_group}
\alias{proc_dist_group}
\title{Procrustes distance between groups}
\usage{
proc_dist_group(specimen_info_file, slider_file, landmarks_file)
}
\arguments{
\item{specimen_info_file}{A data frame with information on the group assignment of specimens in a column named "group".}

\item{slider_file}{A configuration file for semi-sliding landmarks.}

\item{landmarks_file}{A TPS file with landmark coordinates.}
}
\value{
A matrix with inter-centroid distances; A neighbour joining tree.
}
\description{
Takes as arguments a specimen information file, with a grouping column labelled "group"; a sliders file, and a .TPS file. It runs Generalized Procrustes Analysis using geomorph(), and calculates the concensus shape for each group. Then it plots a neighbour joining tree based on the Euclidean distances between the concensus shapes.
}
\examples{
vole_distances <- proc_dist_group("vole_specinfo.csv", "vole_sliders.csv", "vole_lmarks.csv")
}
