% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RunSim}
\alias{RunSim}
\title{RunSim}
\usage{
RunSim(param, regions, doctorVisitingProb, d, s, startWeek,
  outputFile = tempfile(), verbose = F)
}
\arguments{
\item{param}{a}

\item{regions}{a}

\item{doctorVisitingProb}{a}

\item{d}{a}

\item{s}{a}

\item{startWeek}{a}

\item{outputFile}{a}
}
\description{
RunSim
}
