% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SetupStartInfected}
\alias{SetupStartInfected}
\title{SetupStartInfected}
\usage{
SetupStartInfected(betaFile, betas, d, s, startWeek)
}
\arguments{
\item{betaFile}{a}

\item{betas}{a}

\item{d}{a}

\item{s}{a}

\item{startWeek}{a}
}
\description{
SetupStartInfected
}
