% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{single_entity_fake_commuters_2017}
\alias{single_entity_fake_commuters_2017}
\title{A fake commuter dataset for Norway as a single entity in 2017}
\format{\describe{
\item{from}{Location code.}
\item{to}{Location name.}
\item{n}{Number of people.}
}}
\usage{
single_entity_fake_commuters_2017
}
\description{
For use with single_entity_seiiar_2017
}
\keyword{datasets}
