% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{start_points_oslo}
\alias{start_points_oslo}
\title{Names of areas in Norway that previously existed.}
\format{\describe{
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
start_points_oslo
}
\description{
Names of areas in Norway that previously existed.
}
\keyword{datasets}
