% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_group_distances.R
\name{within_group_distances}
\alias{within_group_distances}
\title{group' factor and associated principle component scores;
bs' = number of bootstrap replications.
Distance from origin}
\usage{
within_group_distances(data, bs)
}
\arguments{
\item{data}{A data frame with specimens in rows; first column with group assignment; next columns with ordination scores.}

\item{bs}{Number of bootstrap samples}
}
\value{
A box plot comparing the bootstrapped estimates of the median distance between specimens of each group; a one-way ANOVA result.
}
\description{
Returns the median distance between specimens within each group. The statistic is estimated by sampling with replacement 80% of the observations.
}
\examples{
medians <- within_group_distances(spec_info, 100)
}
