% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_process.R
\name{branching_process}
\alias{branching_process}
\title{Samples from a branching process model with a constant reproduction number}
\usage{
branching_process(
  initial_cases = 1,
  R0 = 3,
  dispersion = 1,
  serial_interval = distcrete::distcrete("gamma", interval = 1, shape =
    epitrix::gamma_mucv2shapescale(mu = 8.4, cv = 3.4/8.4)$shape, scale =
    epitrix::gamma_mucv2shapescale(mu = 8.4, cv = 3.4/8.4)$scale, w = 0),
  days_simulation = 10,
  simulations = 1000
)
}
\arguments{
\item{initial_cases}{Number of initial cases}

\item{R0}{R0}

\item{dispersion}{Dispersion parameter}

\item{serial_interval}{Serial interval (distribution in days)}

\item{days_simulation}{Days simulated}

\item{simulations}{Number of simulations}
}
\description{
R0 - basic reproduction number
serial_interval a distcrete object
}
