% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuter.R
\name{commuter}
\alias{commuter}
\title{commuter}
\usage{
commuter(pop_wo_com = spread::norway_pop_wo_com_2017,
  di_edge_list = spread::norway_di_edge_list_2017,
  start_points = spread::start_points_oslo, r0 = NULL, beta = NULL,
  latent_period = 1.9, infectious_period = 3, asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0, N = 1, M = 7 * 8)
}
\arguments{
\item{pop_wo_com}{Data frame}

\item{di_edge_list}{Data frame}

\item{start_points}{Data frame}

\item{r0}{Float, basic reproduction number}

\item{beta}{Float, infection parameter, 0.6}

\item{latent_period}{Float, 1.9}

\item{infectious_period}{Float, 3}

\item{asymptomatic_prob}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomatic_relative_infectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of repetitions}

\item{M}{Int, Number of days}
}
\description{
commuter
}
