% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuter.R
\name{commuter}
\alias{commuter}
\title{commuter}
\usage{
commuter(seiiar = spread::norway_commuters_2017,
  commuters = spread::norway_seiiar_oslo_2017, r0 = NULL,
  beta = NULL, latent_period = 1.9, infectious_period = 3,
  asymptomatic_prob = 0, asymptomatic_relative_infectiousness = 0,
  N = 1, M = 7 * 8)
}
\arguments{
\item{seiiar}{Data frame}

\item{commuters}{Data frame}

\item{r0}{Float, basic reproduction number}

\item{beta}{Float, infection parameter, 0.6}

\item{latent_period}{Float, 1.9}

\item{infectious_period}{Float, 3}

\item{asymptomatic_prob}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomatic_relative_infectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of repetitions}

\item{M}{Int, Number of days}
}
\description{
commuter
}
