% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuter.R
\name{commuter}
\alias{commuter}
\title{commuter}
\usage{
commuter(seiiar = spread::norway_seiiar_oslo_2017,
  commuters = spread::norway_commuters_2017, r0 = NULL, beta = NULL,
  latent_period = 1.9, infectious_period = 3, asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0, N = 1, M = 7 * 8)
}
\arguments{
\item{seiiar}{Data frame}

\item{commuters}{Data frame}

\item{r0}{Float, basic reproduction number}

\item{beta}{Float, infection parameter, 0.6}

\item{latent_period}{Float, 1.9}

\item{infectious_period}{Float, 3}

\item{asymptomatic_prob}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomatic_relative_infectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{N}{Int = 1 int, Number of repetitions}

\item{M}{Int, Number of days}
}
\description{
This model is a stochastic SEIIaR (susceptible, exposed, infectious, infectious asymptomatic, recovered)
metapopulation model. Each location has a local infection system, while the locations are connected
by people who commute each day. The model differentiates between day and night. During the day you
can infect/be infected in the location where you work, while during the night you can infect/be infected
in the location where you live. It is the same commuters who travel back and forth each day. At the
start of a day, all commuters are sent to their work location, where they mix for 12 hours. The
commuters are then sent to their respective home locations, where they mix for 12 hours. The model
is based upon a published model.
}
\details{
For more information, look at \code{vignette("commuter_model","spread")}
}
