% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commuter.R
\name{commuter}
\alias{commuter}
\title{commuter}
\usage{
commuter(
  seiiar = spread::norway_seiiar_oslo_2017_b2020,
  commuters = spread::norway_commuters_2017_b2020,
  r0 = NULL,
  beta = NULL,
  latent_period = 1.9,
  infectious_period = 3,
  asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0,
  days_simulation = 7 * 8,
  N = 1,
  simulations = 1,
  verbose = TRUE,
  seed = NULL,
  aggregate_location = FALSE
)
}
\arguments{
\item{seiiar}{Data frame containing `location_code`, `S`, `E`, `I`, `Ia`, and `R` for the entire population}

\item{commuters}{Data frame comtaining `from`, `to`, `n` for the number of people who travel}

\item{r0}{Float, basic reproduction number}

\item{beta}{Float, infection parameter, 0.6}

\item{latent_period}{Float, 1.9}

\item{infectious_period}{Float, 3}

\item{asymptomatic_prob}{Float, Proportion/probability of asymptomatic given infectious}

\item{asymptomatic_relative_infectiousness}{Float, Relative infectiousness of asymptomatic infectious}

\item{days_simulation}{Int, Number of days to simulate}

\item{N}{Int = 1 int, Number of internal simulations (average taken). This is generally used for parameter fitting.}

\item{simulations}{Int, Number of simulations (all returned). This is generally used to generate stochastic results.}

\item{verbose}{boolean}

\item{seed}{seed}

\item{aggregate_location}{Do you want to aggregate over all location_codes?}
}
\value{
\describe{
\item{year}{The middle year of a 5 year range (e.g. 2011 is the average of data from 2009-2013).}
\item{location_code}{The location code.}
\item{age}{The population age.}
\item{vax}{The vaccine.}
\item{proportion}{Proportion of people who are vaccinated.}
\item{imputed}{FALSE if real data. TRUE if it is the national average.}
}
}
\description{
This model is a stochastic SEIIaR (susceptible, exposed, infectious, infectious asymptomatic, recovered)
metapopulation model that including commuting.
}
\details{
Each location has a local infection system, while the locations are connected
by people who commute each day. The model differentiates between day and night. During the day you
can infect/be infected in the location where you work, while during the night you can infect/be infected
in the location where you live. It is the same commuters who travel back and forth each day. At the
start of a day, all commuters are sent to their work location, where they mix for 12 hours. The
commuters are then sent to their respective home locations, where they mix for 12 hours.

The model is loosely based upon a published model by Engebretsen (2019) \doi{10.1371/journal.pcbi.1006879}.

For more information, look at \code{vignette("commuter_model","spread")}
}
\examples{
spread::commuter(
  seiiar = spread::norway_seiiar_measles_oslo_2017_b2020,
  commuters = spread::norway_commuters_2017_b2020,
  r0 = 14,
  latent_period = 8,
  infectious_period = 5,
  asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0,
  days_simulation = 7 * 9,
  N = 1,
  simulations = 1,
  verbose = TRUE,
  seed = 4
)
}
