% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_files.R
\docType{data}
\name{norway_seiiar_oslo_2017}
\alias{norway_seiiar_oslo_2017}
\title{SEIIaR data.frame for Norway with 10 people infected in Oslo and everyone susceptible.}
\format{\describe{
\item{location_code}{Location code.}
\item{S}{Number of susceptible people.}
\item{E}{Number of exposed people.}
\item{I}{Number of infectious and symptomatic people.}
\item{Ia}{Number of infectious and asymptomatic people.}
\item{R}{Number of recovered people.}
}}
\usage{
norway_seiiar_oslo_2017
}
\description{
SEIIaR data.frame for Norway with 10 people infected in Oslo and everyone susceptible.
}
\keyword{datasets}
