% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus_shapes.R, R/plot_concensus_shapes.R
\name{plot_concensus_shapes}
\alias{plot_concensus_shapes}
\title{Concensus_shapes}
\usage{
plot_concensus_shapes(
  specimen_info_file,
  slider_file,
  landmarks_file,
  order_vector
)

plot_concensus_shapes(
  specimen_info_file,
  slider_file,
  landmarks_file,
  order_vector
)
}
\arguments{
\item{specimen_info_file:}{a data frame containing at least a specimen group assignment variable, named 'group'.}

\item{slider_file:}{a configuration file for semi-sliding landmarks.}

\item{landmarks_file:}{a TPS file}

\item{order_vector:}{a vector detailing the order of the landmarks for drawing a connecting line.}
}
\value{
returns plots of the mean shape of each group.

returns plots of the mean shape of each group.
}
\description{
Concensus_shapes

plot_concensus_shapes
}
\examples{
plot_concensus_shapes("sheep_specinfo_20230129.csv", "sheep_sliders_202212180739.csv", "sheep_lmrks_202212180740.TPS")
plot_concensus_shapes("sheep_specinfo_20230129.csv", "sheep_sliders_202212180739.csv", "sheep_lmrks_202212180740.TPS")
}
