% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RunSim}
\alias{RunSim}
\title{RunSim}
\usage{
RunSim(id = 1, startVals, R0 = 1.88, gammaTheoretical = 3,
  gammaEffective = gammaTheoretical, a = 1.9, asymptomaticProb = 0.33,
  asymptomaticRelativeInfectiousness = 0.5, M = 100, verbose = F)
}
\arguments{
\item{id}{fileid}

\item{startVals}{a}

\item{R0}{R0}

\item{gammaTheoretical}{days infectious}

\item{gammaEffective}{days infectious}

\item{a}{days latent}

\item{asymptomaticProb}{Probability someone is asymptomatic}

\item{asymptomaticRelativeInfectiousness}{How infectious is this person relative to symptomatic}

\item{M}{days simulated}

\item{verbose}{True/False}
}
\description{
RunSim
}
